requirements.txt
google-genai
pydantic
pydantic-settings
python-dotenv
pytest
requests
notion-client>=2.0.0

# MCP (Model Context Protocol) Integration
# Install with: pip install 'mcp[cli]'
# Or for just the core library: pip install mcp
mcp[cli]>=1.0.0
